V 0.3.3a
DONE
-Everything needed

TODO
-Testing
  -Memory Leaks
  -Actual code is Working
  -Speed performance
