--- Tests/ChrisDataStructuresTests/ChrisDataStructuresTests.swift
@@ -1,11 +1,0 @@
-import XCTest
-import ChrisDataStructures
-
-final class Chris_DataStructuresTests: XCTestCase {
-    
-    func testAddItems() throws {
-        print("Hello World")
-    }
-    
-    
-}
